Import shell aliases
With this patch, we can tangle an `alias.sh` file that, when sourced,
adds useful shell aliases to the current shell.  We source this from
`.bashrc`, so that all interactive (Bash) sessions will be able to use
these aliases.

In particular, we currently have two sets of aliases, which I simply
took from my existing `~/.bash_aliases` file:

  - __`ls` and friends__, which are aliases for making `ls` work a
    little bit nicer, and

  - __`emacsclient`__ aliases, which let me open up a new connection to
    the local running Emacs server on the system either in an existing
    GUI frame, a new GUI frame, or a terminal frame.

I left out a Jekyll alias which I don't find much use for, but I may add
it in later, if I find I want it.

License:       CC0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>