Import and describe prompt configuration
This patch adds a section to tangle a =prompt.sh= file which, when
sourced, sets the Bash shell prompt as I want it.  This is imported from
my current =.bashrc=, but cleaned up and described in much more detail.

I checked both the tangled and weaved output of this section, and the
results are exactly as expected.

License:       CC0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>