Add XDG Base Directory configuration
The XDG Base Directory Specification defines where applications should
look for per-user configuration and data.  By setting these, we can
prevent software from littering our home directory with dotfiles.  For
some software, we will have to configure it specially (normally through
environment variables) to follow this specification (GnuPG, for
instance).  We will do that when we configure that software.

This patch adds a section that, when tangled and sourced, explicitly
sets the variables to the defaults (done for software that doesn't
properly support the defaults).

License:       CC0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>