Import environment variables
This patch adds several more sorts of environment variable
configurations in addition to the XDG Basedir variables we already had
set.  The environment variables were simply brought in from my existing
`.profile` and `.bashrc` files and organized, documented, and
formatted.

When installing per-user software, I tend to put it under `~/.local`.
For this to work, though, we have to set many environment
variables (`$PATH`, for instance, but also `$CFLAGS`, and several other
`$PATH`-like or `$CFLAGS`-like variables).  This patch adds a section
that, when tangled, sets the environment variables properly.

This patch also adds environment variables to use Fcitx as our input
method and Emacsclient as our `$EDITOR`.  At some point in the future,
we'll want to split out the Emacs and Fcitx environment variables into
separate files, for systems other than Italic and Greek, which have X11
and an always-running Emacs daemon.  That way, the combination of our
conditional sourcing with `-r` and installation with GNU Stow will
automatically do the right thing when configuring the user session on
other systems.  For now, I'm not too worried about this.

License:       CC0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>