Teach apparmor about our dotfile setup
The default apparmor configuration in Debian prevents `msmtp` from
reading our configuration file through a symlink.  Turning on apparmor
revealed this problem.  This patch solves this problem by adding a
local override to the apparmor configuration that allows reading any
directory within the user’s home that ends in `msmtp/config`.  This is
probably safe, as such a path is unlikely to be anything but the
directory we want.  We don’t want to hardcode the directory name, as
we could move this git repo somewhere else in the future.

License:       CC0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>